+++ test/test104.cpp
@@ -49,7 +49,7 @@ int main(int argc, char *argv[]) {
 		/**<Update the connectivity and write the para_tree.*/
 		iter = 0;
 		pablo104.updateConnectivity();
-		pablo104.writeTest("Pablo104_iter"+to_string(iter), oct_data);
+		pablo104.writeTest("Pablo104_iter"+to_string(static_cast<unsigned long long>(iter)), oct_data);
 
 		/**<Smoothing iterations on initial data*/
 		int start = 1;
@@ -93,7 +93,7 @@ int main(int argc, char *argv[]) {
 
 			/**<Update the connectivity and write the para_tree.*/
 			pablo104.updateConnectivity();
-			pablo104.writeTest("Pablo104_iter"+to_string(iter), oct_data_smooth);
+			pablo104.writeTest("Pablo104_iter"+to_string(static_cast<unsigned long long>(iter)), oct_data_smooth);
 
 			oct_data = oct_data_smooth;
 		}
