+++ test/test116.cpp
@@ -62,7 +62,7 @@ int main(int argc, char *argv[]) {
 		/**<Update the connectivity and write the para_tree.*/
 		iter = 0;
 		pablo116.updateConnectivity();
-		pablo116.writeTest("Pablo116_iter"+to_string(iter), oct_data);
+		pablo116.writeTest("Pablo116_iter"+to_string(static_cast<unsigned long long>(iter)), oct_data);
 
 		/**<Adapt two times with data injection on new octants.*/
 		int start = 1;
@@ -116,7 +116,7 @@ int main(int argc, char *argv[]) {
 
 			/**<Update the connectivity and write the para_tree.*/
 			pablo116.updateConnectivity();
-			pablo116.writeTest("Pablo116_iter"+to_string(iter), oct_data_new);
+			pablo116.writeTest("Pablo116_iter"+to_string(static_cast<unsigned long long>(iter)), oct_data_new);
 
 			oct_data = oct_data_new;
 		}
@@ -131,7 +131,7 @@ int main(int argc, char *argv[]) {
 
 		/**<Update the connectivity and write the para_tree.*/
 		pablo116.updateConnectivity();
-		pablo116.writeTest("Pablo116_iter"+to_string(iter), oct_data);
+		pablo116.writeTest("Pablo116_iter"+to_string(static_cast<unsigned long long>(iter)), oct_data);
 
 #if NOMPI==0
 	}
