+++ test/test15.cpp
@@ -57,7 +57,7 @@ int main(int argc, char *argv[]) {
 		/**<Update the connectivity and write the para_tree.*/
 		iter = 0;
 		pablo15.updateConnectivity();
-		pablo15.writeTest("Pablo15_iter"+to_string(iter), oct_data);
+		pablo15.writeTest("Pablo15_iter"+to_string(static_cast<unsigned long long>(iter)), oct_data);
 
 		/**<Adapt two times with data injection on new octants.*/
 		int start = 1;
@@ -111,7 +111,7 @@ int main(int argc, char *argv[]) {
 
 			/**<Update the connectivity and write the para_tree.*/
 			pablo15.updateConnectivity();
-			pablo15.writeTest("Pablo15_iter"+to_string(iter), oct_data_new);
+			pablo15.writeTest("Pablo15_iter"+to_string(static_cast<unsigned long long>(iter)), oct_data_new);
 
 			oct_data = oct_data_new;
 		}
@@ -124,7 +124,7 @@ int main(int argc, char *argv[]) {
 
 		/**<Update the connectivity and write the para_tree.*/
 		pablo15.updateConnectivity();
-		pablo15.writeTest("Pablo15_iter"+to_string(iter), oct_data);
+		pablo15.writeTest("Pablo15_iter"+to_string(static_cast<unsigned long long>(iter)), oct_data);
 
 #if NOMPI==0
 	}
