+++ test/test5.cpp
@@ -53,7 +53,7 @@ int main(int argc, char *argv[]) {
 		/**<Update the connectivity and write the para_tree.*/
 		iter = 0;
 		pablo5.updateConnectivity();
-		pablo5.writeTest("Pablo5_iter"+to_string(iter), oct_data);
+		pablo5.writeTest("Pablo5_iter"+to_string(static_cast<unsigned long long>(iter)), oct_data);
 
 		/**<Adapt two times with data injection on new octants.*/
 		int start = 1;
@@ -106,7 +106,7 @@ int main(int argc, char *argv[]) {
 
 			/**<Update the connectivity and write the para_tree.*/
 			pablo5.updateConnectivity();
-			pablo5.writeTest("Pablo5_iter"+to_string(iter), oct_data_new);
+			pablo5.writeTest("Pablo5_iter"+to_string(static_cast<unsigned long long>(iter)), oct_data_new);
 
 			oct_data = oct_data_new;
 		}
